// client/src/context/types.ts
export enum AlertTypes {
  SET_ALERT = 'SET_ALERT',
  REMOVE_ALERT = 'REMOVE_ALERT',
}

export enum AppTypes {
  LOAD_MATRIX = 'LOAD_MATRIX',
  SAVE_MATRIX = 'SAVE_MATRIX',
  ADD_ATTRIBUTE = 'ADD_ATTRIBUTE',
  REMOVE_ATTRIBUTE = 'REMOVE_ATTRIBUTE',
  ADD_COMPANY = 'ADD_COMPANY',
  REMOVE_COMPANY = 'REMOVE_COMPANY',
  SET_RELATION = 'SET_RELATION',
  REMOVE_RELATION = 'REMOVE_RELATION',
  CLEAR_RELATIONS = 'CLEAR_RELATIONS',
  SET_EDITING_ID = 'SET_EDITING_ID',
  SET_NAME = 'SET_NAME',
}

export enum ModalTypes {
  OPEN_MODAL = 'OPEN_MODAL',
  CHANGE_MODAL_VIEW = 'CHANGE_MODAL_VIEW',
  CLOSE_MODAL = 'CLOSE_MODAL',
}